simulators/iot/pumpjack/requirements.txt
click==8.1.7
paho-mqtt==1.6.1
